+++ src/commands/query/handlers/handlers.ts
@@ -5,7 +5,7 @@ import { OutputFormat } from "../../../shared";
 import { getTimeframe } from "../../../services/timeframes/timeframes";
 import dayjs from "dayjs";
 import chalk from "chalk";
-import { NamespaceCombination, QueryFilter } from "../../../services/api/paths/queries";
+import { QueryFilter } from "../../../services/api/paths/queries";
 const utc = require('dayjs/plugin/utc')
 dayjs.extend(utc)
 
@@ -18,8 +18,6 @@ async function createRun(data: {
   from: string,
   to: string,
   id: string,
-  namespaces: string[],
-  combination: NamespaceCombination,
   follow: boolean,
   service: string,
   matchCase: boolean,
@@ -32,8 +30,6 @@ async function createRun(data: {
     needle,
     from,
     to,
-    namespaces,
-    combination,
     follow,
     matchCase,
     regex,